Check output file exists before delete it.
`FileNotFoundError` will be raised when `onefile` command execute failed. At that time, the `onefile_output_filename` may not exists, so add a check before delete it.

+++ nuitka/freezer/Onefile.py
@@ -195,7 +195,8 @@ Categories=Utility;"""
 
     if result != 0:
         # Useless now.
-        os.unlink(onefile_output_filename)
+        if os.path.exists(onefile_output_filename):
+            os.unlink(onefile_output_filename)
 
         stderr = getFileContents(stderr_filename, mode="rb")
